seekBusiness: Migrate to updated visual language
