+++ .changeset/ripe-bananas-invent.md
@@ -1,0 +1,13 @@
+---
+'braid-design-system': minor
+---
+
+---
+updated:
+  - seekBusiness
+---
+
+**seekBusiness:** Migrate to updated visual language
+
+Migrate `seekBusiness` theme to new visual language.
+Adopts the `seekJobs` theme for the latest design standards, rather than the legacy `apac` theme, while retaining the `seekBusiness` brand accent colour.
