+++ packages/braid-design-system/package.json
@@ -193,7 +193,6 @@
     "gradient-parser": "^1.0.0",
     "is-mobile": "^4.0.0",
     "lodash.curry": "^4.1.1",
-    "lodash.merge": "^4.6.2",
     "polished": "^4.1.0",
     "react-focus-lock": "^2.9.7",
     "react-is": "^18.2.0",
@@ -217,7 +216,6 @@
     "@types/autosuggest-highlight": "^3.1.1",
     "@types/babel__core": "^7.20.1",
     "@types/lodash.curry": "^4.1.9",
-    "@types/lodash.merge": "^4.6.9",
     "@types/node": "^18.19.68",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.0",
