+++ packages/braid-design-system/src/lib/themes/baseTokens/nvl.ts
@@ -1,67 +1,51 @@
-import robotoMetrics from '@capsizecss/metrics/roboto';
-import merge from 'lodash.merge';
 import { darken, lighten, rgba, saturate } from 'polished';
-import type { DeepPartial } from 'utility-types';
 
 import { palette } from '../../color/palette';
-import { type BraidTokens, extractFontMetricsForTheme } from '../tokenType';
+import type { BraidTokens } from '../tokenType';
 
 interface MakeTokensOptions {
   name: string;
   displayName: string;
-  brand: string;
   brandAccent: string;
   brandAccentLight: string;
-  brandAccentActive: string;
-  brandAccentHover: string;
   brandAccentSoft: string;
-  brandAccentSoftActive: string;
-  brandAccentSoftHover: string;
-  tokenOverrides?: DeepPartial<BraidTokens>;
 }
 
-const legacyViolet = {
-  900: '#1E0B65',
-  800: '#341B87',
-  700: '#512EAA',
-  600: '#6C41CE',
-  500: '#8B5CEB',
-  400: '#AD84F2',
-  300: '#C6ACF5',
-  200: '#E1D1F9',
-  100: '#F1E8FD',
-  50: '#FAF5FE',
-};
+const formAccent = palette.seekBlueLight['700'];
+const promote = palette.purple['700'];
+const critical = palette.red['700'];
+const focus = rgba(palette.seekBlueLight['300'], 0.7);
+const white = '#fff';
 
 export const makeTokens = ({
   name,
   displayName,
-  brand,
   brandAccent,
   brandAccentLight,
-  brandAccentActive,
-  brandAccentHover,
   brandAccentSoft,
-  brandAccentSoftActive,
-  brandAccentSoftHover,
-  tokenOverrides = {},
 }: MakeTokensOptions): BraidTokens => {
-  const formAccent = palette.indigo['500'];
-  const critical = palette.red['700'];
-  const focus = rgba(palette.indigo['300'], 0.7);
-  const white = '#fff';
+  const brandAccentActive = darken(0.05, brandAccent);
+  const brandAccentHover = lighten(0.05, brandAccent);
+  const brandAccentSoftActive = darken(0.05, brandAccentSoft);
+  const brandAccentSoftHover = darken(0.025, brandAccentSoft);
 
   const tokens: BraidTokens = {
     name,
     displayName,
-    legacy: true,
+    legacy: false,
     typography: {
-      fontFamily: `Roboto, "Helvetica Neue", HelveticaNeue, Helvetica, Arial, sans-serif`,
-      webFont: null,
-      fontMetrics: extractFontMetricsForTheme(robotoMetrics),
+      fontFamily: 'SeekSans, "SeekSans Fallback", Arial, Tahoma, sans-serif',
+      webFont: 'https://www.seek.com.au/static/shared-web/seeksans.css',
+      fontMetrics: {
+        capHeight: 783,
+        ascent: 1057,
+        descent: -274,
+        lineGap: 0,
+        unitsPerEm: 1000,
+      },
       fontWeight: {
         regular: 400,
-        medium: 600,
+        medium: 500,
         strong: 700,
       },
       heading: {
@@ -73,84 +57,84 @@ export const makeTokens = ({
           '1': {
             mobile: {
               fontSize: 28,
-              rows: 9,
+              lineGap: 11,
             },
             tablet: {
-              fontSize: 42,
-              rows: 11,
+              fontSize: 36,
+              lineGap: 14,
             },
           },
           '2': {
             mobile: {
-              fontSize: 21,
-              rows: 8,
+              fontSize: 24,
+              lineGap: 11,
             },
             tablet: {
-              fontSize: 28,
-              rows: 9,
+              fontSize: 30,
+              lineGap: 13,
             },
           },
           '3': {
             mobile: {
-              fontSize: 21,
-              rows: 7,
+              fontSize: 22,
+              lineGap: 10,
             },
             tablet: {
-              fontSize: 21,
-              rows: 7,
+              fontSize: 24,
+              lineGap: 11,
             },
           },
           '4': {
             mobile: {
-              fontSize: 18,
-              rows: 7,
+              fontSize: 20,
+              lineGap: 9,
             },
             tablet: {
-              fontSize: 18,
-              rows: 7,
+              fontSize: 20,
+              lineGap: 9,
             },
           },
         },
       },
       text: {
-        xsmall: {
+        large: {
           mobile: {
-            fontSize: 12,
-            rows: 5,
+            fontSize: 18,
+            lineGap: 13,
           },
           tablet: {
-            fontSize: 12,
-            rows: 5,
-          },
-        },
-        small: {
-          mobile: {
-            fontSize: 14,
-            rows: 5,
-          },
-          tablet: {
-            fontSize: 14,
-            rows: 5,
+            fontSize: 18,
+            lineGap: 13,
           },
         },
         standard: {
           mobile: {
             fontSize: 16,
-            rows: 6,
+            lineGap: 12,
           },
           tablet: {
             fontSize: 16,
-            rows: 6,
+            lineGap: 12,
           },
         },
-        large: {
+        small: {
           mobile: {
-            fontSize: 18,
-            rows: 7,
+            fontSize: 14,
+            lineGap: 10,
           },
           tablet: {
-            fontSize: 18,
-            rows: 7,
+            fontSize: 14,
+            lineGap: 10,
+          },
+        },
+        xsmall: {
+          mobile: {
+            fontSize: 12,
+            lineGap: 9,
+          },
+          tablet: {
+            fontSize: 12,
+            lineGap: 9,
           },
         },
       },
@@ -162,17 +146,17 @@ export const makeTokens = ({
       large: 1280,
     },
     grid: 4,
-    touchableSize: 11,
+    touchableSize: 12,
     space: {
       gutter: 6,
-      xxsmall: 1,
-      xsmall: 2,
-      small: 3,
-      medium: 5,
+      xxsmall: 2,
+      xsmall: 3,
+      small: 4,
+      medium: 6,
       large: 8,
       xlarge: 12,
-      xxlarge: 24,
-      xxxlarge: 30,
+      xxlarge: 16,
+      xxxlarge: 24,
     },
     transforms: {
       touchable: 'scale(0.95)',
@@ -183,14 +167,14 @@ export const makeTokens = ({
     },
     border: {
       radius: {
-        small: '2px',
-        standard: '4px',
-        large: '6px',
-        xlarge: '10px',
+        small: '4px',
+        standard: '8px',
+        large: '16px',
+        xlarge: '24px',
       },
       width: {
-        standard: 1,
-        large: 2,
+        standard: 2,
+        large: 4,
       },
       color: {
         brandAccent,
@@ -202,19 +186,19 @@ export const makeTokens = ({
         field: palette.grey['400'],
         focus,
         formAccent,
-        formAccentLight: palette.indigo['300'],
+        formAccentLight: palette.seekBlueLight['300'],
         info: palette.blue['700'],
         infoLight: palette.blue['300'],
         neutral: palette.grey['700'],
-        neutralLight: palette.grey['200'],
+        neutralLight: palette.grey['100'],
         neutralInverted: white,
         positive: palette.mint['700'],
         positiveLight: palette.mint['300'],
-        promote: legacyViolet['700'],
-        promoteLight: legacyViolet['300'],
+        promote,
+        promoteLight: palette.purple['300'],
       },
     },
-    focusRingSize: 3,
+    focusRingSize: 6,
     shadows: {
       small: [
         `0 2px 4px 0px ${rgba(palette.grey['800'], 0.1)}`,
@@ -241,28 +225,28 @@ export const makeTokens = ({
         critical,
         criticalLight: palette.red['300'],
         formAccent,
-        formAccentLight: palette.indigo['300'],
+        formAccentLight: palette.seekBlueLight['300'],
         info: palette.blue['700'],
         infoLight: palette.blue['300'],
-        link: formAccent,
-        linkLight: palette.indigo['300'],
-        linkHover: formAccent,
-        linkVisited: legacyViolet['700'],
-        linkLightVisited: legacyViolet['300'],
+        link: palette.grey['700'],
+        linkLight: white,
+        linkHover: palette.grey['700'],
+        linkVisited: palette.purple['800'],
+        linkLightVisited: palette.purple['200'],
         neutral: palette.grey['700'],
         neutralInverted: white,
         positive: palette.mint['700'],
         positiveLight: palette.mint['300'],
-        promote: legacyViolet['700'],
-        promoteLight: legacyViolet['300'],
-        rating: palette.seekPink['500'],
+        promote,
+        promoteLight: palette.purple['300'],
+        rating: brandAccent,
         secondary: palette.grey['500'],
         secondaryInverted: rgba('#fff', 0.65),
       },
       background: {
-        body: palette.grey['50'],
-        bodyDark: palette.grey['900'],
-        brand,
+        body: '#fff',
+        bodyDark: palette.grey['800'],
+        brand: palette.seekBlue['700'],
         brandAccent,
         brandAccentActive,
         brandAccentHover,
@@ -280,28 +264,28 @@ export const makeTokens = ({
         criticalSoftHover: darken(0.025, palette.red['50']),
         formAccent,
         formAccentActive: darken(0.05, formAccent),
-        formAccentHover: saturate(0.5, lighten(0.075, formAccent)),
-        formAccentSoft: palette.indigo['50'],
-        formAccentSoftActive: darken(0.05, palette.indigo['50']),
-        formAccentSoftHover: darken(0.025, palette.indigo['50']),
+        formAccentHover: lighten(0.075, formAccent),
+        formAccentSoft: palette.seekBlueLight['50'],
+        formAccentSoftActive: darken(0.05, palette.seekBlueLight['50']),
+        formAccentSoftHover: darken(0.025, palette.seekBlueLight['50']),
         info: palette.blue['700'],
         infoLight: palette.blue['100'],
         neutral: palette.grey['700'],
         neutralActive: darken(0.05, palette.grey['700']),
         neutralHover: lighten(0.05, palette.grey['700']),
-        neutralLight: palette.grey['100'],
-        neutralSoft: palette.grey['50'],
-        neutralSoftActive: darken(0.05, palette.grey['50']),
-        neutralSoftHover: darken(0.025, palette.grey['50']),
+        neutralLight: palette.grey['75'],
+        neutralSoft: palette.grey['75'],
+        neutralSoftActive: darken(0.05, palette.grey['75']),
+        neutralSoftHover: darken(0.025, palette.grey['75']),
         positive: palette.mint['700'],
         positiveLight: palette.mint['100'],
-        promote: legacyViolet['700'],
-        promoteLight: legacyViolet['100'],
+        promote,
+        promoteLight: palette.purple['100'],
         surface: white,
         surfaceDark: palette.grey['800'],
       },
     },
   };
 
-  return merge(tokens, tokenOverrides);
+  return tokens;
 };
